SELinux weirdness needs better guards
At first, before firewalld is installed, there's a missing python
binding for selinux so the conditions weren't met. After second run (for
idempotence checks, among other things) it wasn't missing but the
selinux object had an attribute "status" equal to "disabled". Fixes
needed to be made since it isn't a straight-up boolean afterr all, it's
an object with attributes that will, on its face, always be truthy.