Molecule tests only should skip Python compile slowdown
It works. We know it works. If you don't believe me, remove the tags and
run it once or twice. It's even idempotent. I just don't want to wait
for the compile step time and time again when configuring the rest of
the OS. Those are the actually brittle steps.